isolate a MapTooltip component
